Don't delay logging until the end of test when GRPC_ALWAYS_LOG is enabled
Motivation:

In the tests we capture all emitted logs and print them if a test fails,
or if the "GRPC_ALWAYS_LOG" environment variable is set. However, in the
latter case we will only print the logs if test execution completes; we
get no logs in the case of wedged or crashing tests.

Modifications:

- Extract log formatting into 'CapturedLogFormatter'
- Add an option to the 'CapturingLogHandlerFactory' to allow logs to be
  print when captured
- Only emit logs at the end of test if 'GRPC_ALWAYS_LOG' is disabled and
  the test fails.

Result:

When 'GRPC_ALWAYS_LOG' is enabled, logs will be printed as they are
caught.
